Move code into a central component for Android
